requirements.txt
asgiref==3.8.1
Django==4.2
djangorestframework==3.15.1
isort==5.13.2
psycopg==3.1.19
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-slugify==8.0.4
sqlparse==0.5.0
text-unidecode==1.3
typing_extensions==4.12.0
